Fix warning about bool XOR int

+++ source/tcppLibrary.hpp
@@ -1581,7 +1581,7 @@ namespace tcpp
 
 			// even number of NOTs: false ^ false = false, false ^ true = true
 			// odd number of NOTs: true ^ false = true (!false), true ^ true = false (!true)
-			return resultApply ^ evalPrimary();
+			return static_cast<int>(resultApply) ^ evalPrimary();
 		};
 
 		auto evalMultiplication = [&tokens, &evalUnary]()
